Fix more things
