+++ src/common.rs
@@ -52,5 +52,4 @@ pub(crate) use ::{
 pub(crate) use ::{
   std::{future::Future, time::Duration},
   tempfile::TempDir,
-  tokio::io::{AsyncReadExt, AsyncWriteExt},
 };
